relocated public C and C++ headers
