+++ include/ocilibc/config.h
